+++ src/tests/EntityGraphQL.Tests/MutationTests/MutationTests.cs
@@ -971,5 +971,15 @@ namespace EntityGraphQL.Tests
             Assert.Null(results.Errors);
             Assert.Equal(0, results.Data["getValue"]);
         }
+
+        [Fact]
+        public void TestNoArgsOnInputType()
+        {
+            var schema = SchemaBuilder.FromObject<TestDataContext>();
+            schema.AddInputType<InputObject>("InputObject", "Using an object in the arguments");
+
+            var ex = Assert.Throws<EntityQuerySchemaException>(() => schema.Type<InputObject>().AddField("invalid", new { id = (int?)null }, (ctx, args) => 8, "Invalid field"));
+            Assert.Equal("Field invalid on type InputObject has arguments but is a GraphQL Input type and can not have arguments.", ex.Message);
+        }
     }
 }